+++ src/pages/login/EmailCheckPage.tsx
@@ -4,9 +4,9 @@ import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
 
-import TermsPage from './Terms/TermsPage';
+import TermsPage from './TermsCheckPage/TermsPage';
 
-import s from "./EmailCheckPage.module.scss"
+import s from "./styles/EmailCheckPage.module.scss"
 
 import kakaoTalkIconWithBackground from "../../assets/loginPage/kakakTalkIconWithBackground.svg";
 import leftBraket from "../../assets/leftBraket.svg";
