+++ src/pages/login/TermsCheckPage/assets/Terms.ts
