+++ src/pages/login/TermsCheckPage/assets/TermsCheck.ts
