+++ src/pages/login/TermsCheckPage/assets/TermsData.ts
