+++ src/pages/login/styles/EmailCheckPage.module.scss
